polish form class created
