+++ MyCalculator/app/src/main/java/com/example/totha/mycalculator/MainActivity.java
@@ -30,7 +30,6 @@ import java.util.Stack;
 import java.util.TreeSet;
 
 
-
 /*          TO_DO           */
 /*  - AUTOMATIC SRCOLL DOWN
     - GRIDLAYOUT SIZE ELEMNTS
@@ -38,6 +37,7 @@ import java.util.TreeSet;
 
 public class MainActivity extends AppCompatActivity implements View.OnClickListener{
 
+    PolishFormAlgorithm polishForm;
     TextView resultTextView;
     StringBuilder rawStringBuilder;
     Map<String, Double> resultsMap;
@@ -58,6 +58,9 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         setContentView(R.layout.activity_main);
         resultTextView = (TextView) findViewById(R.id.resultTextView);
         GridLayout buttonGridLayout = (GridLayout) findViewById(R.id.buttonGridLayout);
+
+        polishForm = new PolishFormAlgorithm();
+
         resultsMap = new HashMap<String, Double>();
 
 
@@ -95,7 +98,7 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         switch(buttonText) {
             case "=":
                 if(Character.isDigit(finalString.charAt(finalString.length() - 1))){
-                    double res = reversePolishForm(polishForm(finalString));
+                    double res = polishForm.reversePolishForm(polishForm.toPolishForm(finalString));
                     resultTextView.setText(finalString + " = " + res);
                     rawStringBuilder.setLength(0);
                     rawStringBuilder.append(res);
@@ -187,47 +190,6 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
         }
     }
 
-    /**
-     * Called by onClick method if user press "=" button
-     * Returns a double that calculated from an equation
-     * it will be wrote to the reult textview
-     * @param components contains the operators and numbers in a row
-     * @return calculated result
-     */
-    public double reversePolishForm(ArrayList<String> components) {
-        if(components.isEmpty()) return 0;
-        Stack<Double> stack = new Stack<Double>();;
-        double num1;
-        double num2;
-
-        for(String comp : components){
-            if(comp.equals("+")) {
-                num2 = stack.pop();
-                num1 = stack.pop();
-                stack.push(num1 + num2);
-            }
-            else if(comp.equals("-")) {
-                num2 = stack.pop();
-                num1 = stack.pop();
-                stack.push(num1 - num2);
-            }
-            else if(comp.equals("*")) {
-                num2 = stack.pop();
-                num1 = stack.pop();
-                stack.push(num1 * num2);
-            }
-            else if(comp.equals("/")){
-                num2 = stack.pop();
-                num1 = stack.pop();
-                stack.push(num1 / num2);
-            }
-            else {
-                stack.push(Double.parseDouble(comp));
-            }
-         }
-        return stack.pop();
-    }
-
 
     /*
     private void showAndSaveNewInput(String newInput){
@@ -254,56 +216,6 @@ public class MainActivity extends AppCompatActivity implements View.OnClickListe
     }
     */
 
-
-
-    /**
-     * Returns an ArrayList object. The equationString argument
-     * is the equation in the normal form. The polishForm method shape the
-     * equation in polish form and save the arguments (in String form) in ArrayList
-     * @param equationString is the equation in normal form
-     * @return an ArrayList
-     */
-    public ArrayList<String> polishForm(String equationString){
-        int i = 0;
-        ArrayList<String> polishF = new ArrayList<String>();
-        Stack<Character> operators = new Stack<Character>();
-        String number = "";
-        while(i < equationString.length()){
-            Character character = equationString.charAt(i);
-            if(character == '+' || character == '-'){
-                polishF.add(number);
-                number = "";
-                if(operators.isEmpty()) operators.push(character);
-                else{
-                    while(!operators.isEmpty()){
-                        Character stackCharacter = operators.pop();
-                        polishF.add(stackCharacter.toString());
-                    }
-                    operators.push(character);
-                }
-            } else if(character == '*' || character == '/'){
-                polishF.add(number);
-                number = "";
-                if(operators.isEmpty()) operators.push(character);
-                else{
-                    while(!operators.isEmpty() && (operators.peek() != '+' && operators.peek() != '-')){
-                        Character stackCharacter = operators.pop();
-                        polishF.add(stackCharacter.toString());
-                    }
-                    operators.push(character);
-                }
-            } else{
-                number += character;
-            }
-            ++i;
-        }
-        if(!number.equals("")) polishF.add(number);
-        while(!operators.isEmpty()){
-            Character stackCharacter = operators.pop();
-            polishF.add(stackCharacter.toString());
-        }
-        return polishF;
-    }
