+++ MyCalculator/app/src/main/java/com/example/totha/mycalculator/PolishFormInterface.java
@@ -1,0 +1,11 @@
+package com.example.totha.mycalculator;
+
+import java.util.ArrayList;
+
+/**
+ * Created by totha on 2016. 07. 21..
+ */
+public interface PolishFormInterface {
+    public double reversePolishForm(ArrayList<String> components);
+    public ArrayList<String> toPolishForm(String equationString);
+}
