+++ MyCalculator/app/src/main/res/layout/activity_main.xml
@@ -58,12 +58,7 @@
         android:columnCount="4"
         android:rowCount="4"
         android:orientation="horizontal"
-        android:useDefaultMargins="false"
-        android:columnOrderPreserved="false"
-        android:transitionGroup="false"
-        android:touchscreenBlocksFocus="false"
-        android:layout_margin="0dp"
-        android:padding="0dp">
+        android:useDefaultMargins="false">
 
             <Button
                 android:layout_width="wrap_content"
